delete useless interfaces
