+++ src/main/java/configurator/Configurator.java
@@ -10,28 +10,96 @@ import parser.Parser;
 import reverser.Reverser;
 import temp_files_fabric.TempFilesFabric;
 
+import java.io.File;
 import java.util.Comparator;
 
-public interface Configurator<T> {
-    void config(Comparator<T> comparator, Parser<T> parser);
+public class Configurator<T> {
+    private Comparator<T> comparator;
+    private TempFilesFabric tempFilesFabric = new TempFilesFabric();
+    private FileMerger<T> fileMerger = new FileMerger<>();
+    private FileDivider fileDivider = new FileDivider();
+    private Parser<T> parser;
+    private FileSorter<T> fileSorter = new FileSorter<>();
+    private ListMergeSorter<T> listMergeSorter = new ListMergeSorter<>();
+    private Checker<T> checker = new Checker<>();
+    private Manager<T> manager = new Manager<>();
+    private Reverser reverser = new Reverser();
 
-    Comparator<T> getComparator();
+    public void config(Comparator<T> comparator, Parser<T> parser) {
+        this.comparator = comparator;
+        this.parser = parser;
 
-    TempFilesFabric getTempFilesFabric();
+        tempFilesFabric.setTempDirectory(new File("./tmp/"));
 
-    FileMerger<T> getFileMerger();
+        fileMerger.setComparator(comparator);
+        fileMerger.setParser(parser);
+        fileMerger.setTempFilesFabric(tempFilesFabric);
 
-    FileDivider getFileDivider();
+        fileDivider.setTempFilesFabric(tempFilesFabric);
+        fileDivider.setMaxSize(10_000_000);
 
-    Parser<T> getParser();
+        listMergeSorter.setComparator(comparator);
 
-    FileSorter<T> getFileSorter();
+        checker.setComparator(comparator);
+        checker.setParser(parser);
 
-    ListMergeSorter<T> getListMergeSorter();
+        fileSorter.setComparator(comparator);
+        fileSorter.setFileDivider(fileDivider);
+        fileSorter.setParser(parser);
+        fileSorter.setTempFilesFabric(tempFilesFabric);
+        fileSorter.setListMergeSorter(listMergeSorter);
+        fileSorter.setFileMerger(fileMerger);
+        fileSorter.setSizeLimit(100_000_000);
 
-    Checker<T> getChecker();
+        reverser.setFileDivider(fileDivider);
+        reverser.setTempFilesFabric(tempFilesFabric);
+        reverser.setSizeLimit(100_000_00);
 
-    Manager<T> getManager();
+        manager.setFileMerger(fileMerger);
+        manager.setFileSorter(fileSorter);
+        manager.setChecker(checker);
+        manager.setTempFilesFabric(tempFilesFabric);
+        manager.setReverser(reverser);
 
-    Reverser getReverser();
+    }
+
+    public Comparator<T> getComparator() {
+        return comparator;
+    }
+
+    public TempFilesFabric getTempFilesFabric() {
+        return tempFilesFabric;
+    }
+
+    public FileMerger<T> getFileMerger() {
+        return fileMerger;
+    }
+
+    public FileDivider getFileDivider() {
+        return fileDivider;
+    }
+
+    public Parser<T> getParser() {
+        return parser;
+    }
+
+    public FileSorter<T> getFileSorter() {
+        return fileSorter;
+    }
+
+    public ListMergeSorter<T> getListMergeSorter() {
+        return listMergeSorter;
+    }
+
+    public Checker<T> getChecker() {
+        return checker;
+    }
+
+    public Manager<T> getManager() {
+        return manager;
+    }
+
+    public Reverser getReverser() {
+        return reverser;
+    }
 }
