+++ src/main/java/temp_files_fabric/TempFilesFabric.java
@@ -2,7 +2,21 @@ package temp_files_fabric;
 
 import java.io.File;
 
-public interface TempFilesFabric {
-    void setTempDirectory(File directory);
-    File getNewTempFile();
+public class TempFilesFabric{
+    private static long tempNumber = 0;
+    private File directory = new File(".");
+
+
+    public void setTempDirectory(File directory) {
+        if (directory.isDirectory()) {
+            this.directory = directory;
+        }
+    }
+
+    public File getNewTempFile() {
+        tempNumber++;
+        File result = new File(directory.getPath() + "/" + "temp" + tempNumber + ".txt");
+        result.deleteOnExit();
+        return result;
+    }
 }
